// src/gen/model/payroll-nz/earningsRates.ts
/**
 * Xero Payroll NZ
 * This is the Xero Payroll API for orgs in the NZ region.
 *
 * The version of the OpenAPI document: 2.3.7
 * Contact: api@xero.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EarningsRate } from '././earningsRate';
import { Pagination } from '././pagination';
import { Problem } from '././problem';

export class EarningsRates {
    'pagination'?: Pagination;
    'problem'?: Problem;
    'earningsRates'?: Array<EarningsRate>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "pagination",
            "baseName": "pagination",
            "type": "Pagination"
        },
        {
            "name": "problem",
            "baseName": "problem",
            "type": "Problem"
        },
        {
            "name": "earningsRates",
            "baseName": "earningsRates",
            "type": "Array<EarningsRate>"
        }    ];

    static getAttributeTypeMap() {
        return EarningsRates.attributeTypeMap;
    }
}